+++ src/fslayer.c
@@ -109,8 +109,8 @@ fslayer_read(int fd, void *buf, size_t count)
 	} else {
 		noloop = 1;
 		ret = read(fd, buf, count);
-		STRACE((strace, "read(%d, buf, %d) = %d\n",
-			fd, (int)count, (int)ret));
+		STRACE((strace, "read(%d, buf, %zu) = %zd\n",
+			fd, count, ret));
 		noloop = 0;
 	}
 	return (ret);
@@ -126,8 +126,8 @@ fslayer_write(int fd, const void *buf, size_t count)
 	} else {
 		noloop = 1;
 		ret = write(fd, buf, count);
-		STRACE((strace, "write(%d, buf, %d) = %d\n",
-			fd, (int)count, (int)ret));
+		STRACE((strace, "write(%d, buf, %zu) = %zd\n",
+			fd, count, ret));
 		noloop = 0;
 	}
 	return (ret);
@@ -143,8 +143,8 @@ fslayer_lseek(int fildes, off_t offset, int whence)
 	} else {
 		noloop = 1;
 		ret = lseek(fildes, offset, whence);
-		STRACE((strace, "lseek(%d, %d, %d) = %d\n",
-		    fildes, (int)offset, whence, (int)ret));
+		STRACE((strace, "lseek(%d, %jd, %d) = %jd\n",
+		    fildes, (intmax_t)offset, whence, (intmax_t)ret));
 		noloop = 0;
 	}
 	return (ret);
@@ -201,8 +201,8 @@ fslayer_lstat(const char *path, struct stat *buf)
 			unless (ret) {
 				fprintf(strace, "mode=%o", buf->st_mode);
 				if (S_ISREG(buf->st_mode)) {
-					fprintf(strace, ", size=%d",
-					    (int)buf->st_size);
+					fprintf(strace, ", size=%jd",
+					    (intmax_t)buf->st_size);
 				}
 			}
 			fprintf(strace, "}) = %d\n", ret);
@@ -233,8 +233,8 @@ fslayer_stat(const char *path, struct stat *buf)
 			unless (ret) {
 				fprintf(strace, "mode=%o", buf->st_mode);
 				if (S_ISREG(buf->st_mode)) {
-					fprintf(strace, ", size=%d",
-					    (int)buf->st_size);
+					fprintf(strace, ", size=%jd",
+					    (intmax_t)buf->st_size);
 				}
 			}
 			fprintf(strace, "}) = %d\n", ret);
@@ -259,8 +259,8 @@ fslayer_fstat(int fd, struct stat *buf)
 			unless (ret) {
 				fprintf(strace, "mode=%o", buf->st_mode);
 				if (S_ISREG(buf->st_mode)) {
-					fprintf(strace, ", size=%d",
-					    (int)buf->st_size);
+					fprintf(strace, ", size=%jd",
+					    (intmax_t)buf->st_size);
 				}
 			}
 			fprintf(strace, "}) = %d\n", ret);
